+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2018-11-04
 ### Added
 - Instructions for running using clisp
 - dont-understand response
@@ -38,5 +40,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - brain replies hardcoded value "Ой, приветик!" to "Привет" message
 - brain replies "unknown" to messages it doesn't understand
 
-[Unreleased]: https://github.com/aragaer/pa_brain/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/aragaer/pa_brain/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/aragaer/pa_brain/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/aragaer/pa_brain/compare/v0.1.0...v0.2.0